Move Motion Menu to its own file
