+++ Marlin/src/lcd/ultralcd.h
@@ -288,6 +288,7 @@
 
     // Manual Movement
     constexpr float manual_feedrate_mm_m[XYZE] = MANUAL_FEEDRATE;
+    extern float move_menu_scale;
     #if IS_KINEMATIC
       extern bool processing_manual_move;
     #else
